refactor(abstract-eth): add common method to abstract-eth WIN-1087 TICKET: WIN-1087
