+++ modules/sdk-coin-arbeth/src/lib/index.ts
@@ -4,3 +4,4 @@ export { TransactionBuilder } from './transactionBuilder';
 export { TransferBuilder } from './transferBuilder';
 export { Transaction, KeyPair } from '@bitgo/abstract-eth';
 export { Utils };
+export * from './walletUtil';
