+++ modules/sdk-coin-opeth/package.json
@@ -43,8 +43,7 @@
     "@bitgo/abstract-eth": "^1.6.0",
     "@bitgo/sdk-core": "^8.26.0",
     "@bitgo/statics": "^29.0.0",
-    "@ethereumjs/common": "^2.6.5",
-    "ethereumjs-abi": "^0.6.5"
+    "@ethereumjs/common": "^2.6.5"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.24.0",
