+++ modules/sdk-coin-opeth/src/lib/transactionBuilder.ts
@@ -1,12 +1,6 @@
 import { BaseCoin as CoinConfig } from '@bitgo/statics';
-import EthereumAbi from 'ethereumjs-abi';
 import { BuildTransactionError, TransactionType } from '@bitgo/sdk-core';
-import {
-  KeyPair,
-  Transaction,
-  TransactionBuilder as EthLikeTransactionBuilder,
-  walletSimpleConstructor,
-} from '@bitgo/abstract-eth';
+import { KeyPair, Transaction, TransactionBuilder as EthLikeTransactionBuilder } from '@bitgo/abstract-eth';
 import { getCommon } from './utils';
 import { TransferBuilder } from './transferBuilder';
 import { walletSimpleByteCode } from './walletUtil';
@@ -18,20 +12,7 @@ export class TransactionBuilder extends EthLikeTransactionBuilder {
     super(_coinConfig);
     this._common = getCommon(this._coinConfig.network.type);
     this.transaction = new Transaction(this._coinConfig, this._common);
-  }
-
-  /**
-   * Returns the smart contract encoded data
-   *
-   * @param {string[]} addresses - the contract signers
-   * @returns {string} - the smart contract encoded data
-   */
-  protected getContractData(addresses: string[]): string {
-    const params = [addresses];
-    const resultEncodedParameters = EthereumAbi.rawEncode(walletSimpleConstructor, params)
-      .toString('hex')
-      .replace('0x', '');
-    return walletSimpleByteCode + resultEncodedParameters;
+    this._walletSimpleByteCode = walletSimpleByteCode;
   }
 
   /** @inheritdoc */
