+++ modules/sdk-coin-polygon/src/lib/index.ts
@@ -4,3 +4,4 @@ export { TransferBuilder } from './transferBuilder';
 
 import * as Utils from './utils';
 export { Utils };
+export * from './walletUtil';
